refactor(consultation): nettoyage post-revue slice 2 — i18n, cycle de type, a11y
Suite à la revue multi-agents de la PR #609 (0 finding ≥ MEDIUM) :

- L1 : purge des clés i18n orphelines du namespace `consultation` (×3 langues)
  laissées par la suppression des onglets bespoke — `tabsLabel`, `tabs.overview/
  glycemia/treatment/documents`, et les objets `overview`/`glycemia`/`treatment`.
  Conserve `tabs.glycemicProfile` + les clés du drawer/TabState.
- L2 : relocalise le type `ContextFlags` dans le module neutre
  `patient-record-views.ts` → casse le cycle de type PatientContextBar ↔
  PatientAlertFlags. PatientContextBar le ré-exporte (rétro-compat ReviewClient).
- a11y (WCAG 2.4.7) : indicateur de focus visible sur le heading du drawer
  (focalisé par programme à l'ouverture) — même outline que les boutons.

HDS-LOW (id/publicRef dans le DTO en mémoire client) = durcissement serveur
hors périmètre, à tracer en suivi. Scrim bg-black/30 (L3) pré-existant, inchangé.
